pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "k8s-yaml-fmt"
version = "0.1.0"
description = "Kubernetes YAML formatter with idiomatic key ordering"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "you@example.com"}
]
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
]
keywords = ["kubernetes", "yaml", "formatter", "k8s", "pre-commit"]
dependencies = [
    "ruamel.yaml>=0.18",
]

[project.scripts]
k8s-yaml-fmt = "k8s_yaml_fmt:main"

[project.urls]
Homepage = "https://github.com/yourusername/k8s-yaml-fmt"
Repository = "https://github.com/yourusername/k8s-yaml-fmt"

[tool.setuptools]
py-modules = ["k8s_yaml_fmt"]

[tool.pytest.ini_options]
testpaths = ["."]
python_files = ["test_*.py"]

[tool.ruff]
line-length = 120
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]
